[fio toup] tests: Make p12 credentials compatible with default OpenSSL 3
Current test credentials use the RC2-40-CBC algorithm for encryption of
the PKCS7 data inside the PKCS12 archive. This commit replaces it with
AES-256-CBC, allowing OpenSSL 3.x to be used without the need of
enabling the legacy provider, while still keeping compatibility with
OpenSSL 1.x.
The keys and certificates themselves were not affected.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>