[fio extras] Reduce severity of log messages during an update
This commit differentiates, based on the client state, the severity of
some log messages shown when the content of snapshot or targets metadata
files do not match the expected hash. During a typical update, when the
hash verification is expected to fail, the messages are logged as info.
Otherwise, they are logged as errors, keeping the current behavior.

Signed-off-by: Andre Detsch <andre.detsch@foundries.io>