[fio extras] config: Allow not setting a default pacman
Enable libaktualizr clients to determine whether to designate the
ostree package manager as the default manager or not.
It may be crucial not to set it if there is a need to distinguish
between cases where the 'ostree' package manager is configured by
a user and cases where a user hasn't configured any package manager.
It requires the libaktualizr client deduce the package manager type
for application. If such a scenario arises, the libaktualizr client
should define 'DONT_DEFINE_DEFAULT_MANAGER' in its build system.

Signed-off-by: Mike Sul <mike.sul@foundries.io>